Introduced abstraction between general Pullrequest views and pullrequest overview view.
