+++ src/test/java/nl/tudelft/ewi/devhub/webtests/views/PullRequestView.java
@@ -4,11 +4,10 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-import static org.junit.Assert.assertTrue;
-
-public class PullRequestView extends AuthenticatedView {
-
+public abstract class PullRequestView extends AuthenticatedView {
+	
 	private static final By BREADCRUMB = By.xpath("/html/body/div/ol");
+	
 	private static final By HEADERS = By.xpath("//div[@class='headers']");
 	private static final By MESSAGE_HEADER = By.xpath(".//h2[@class='header']");
 	private static final By AUTHOR_SUB_HEADER = By.xpath(".//h5[@class='subheader']");
@@ -17,11 +16,6 @@ public class PullRequestView extends AuthenticatedView {
 		super(driver);
 	}
 
-	@Override
-	protected void invariant() {
-		super.invariant();
-		assertTrue(currentPathContains("/pull/"));
-	}
 
 	protected WebElement getBreadcrumb() {
 		return getDriver().findElement(BREADCRUMB);
@@ -58,7 +52,7 @@ public class PullRequestView extends AuthenticatedView {
 	 * Navigate to the course view. Note that the course view currently
 	 * does not exist, and redirects to the course edition page.
 	 * @return a {@link CoursesView}.
-	 * @see PullRequestView#toCourseEditionView()
+	 * @see PullRequestOverViewView#toCourseEditionView()
      */
 	public AuthenticatedView toCourseView() {
 		getBreadcrumb().findElement(By.xpath("./li[2]/a")).click();
@@ -99,35 +93,6 @@ public class PullRequestView extends AuthenticatedView {
      */
 	public String getPullRequestName() {
 		return getBreadcrumb().findElement(By.xpath("./li[6]")).getText();
-	}
-
-	/**
-	 * Check if the pull request is open.
-	 * @return Returns true if the pull request is open.
-     */
-	public boolean isOpen() {
-		return getDriver().findElement(By.id("btn-close")) != null;
-	}
-
-	/**
-	 * Close the pull request.
-	 */
-	public void close() {
-		getDriver().findElement(By.id("btn-close")).click();
-	}
-
-	/**
-	 * Merge the pull request.
-	 */
-	public void merge() {
-		getDriver().findElement(By.id("btn-merge")).click();
-	}
-
-	/**
-	 * Remove the branch.
-	 */
-	public void removeBranch() {
-		getDriver().findElement(By.id("btn-remove-branch")).click();
 	}
 
 }
